Move workspace sidebar preview into app shell
