+++ packages/web-core/src/pages/workspaces/Workspaces.tsx
@@ -1,9 +1,5 @@
 import { WorkspacesLayout } from './WorkspacesLayout';
 
-interface WorkspacesProps {
-  isAppBarHovered?: boolean;
-}
-
-export function Workspaces({ isAppBarHovered = false }: WorkspacesProps) {
-  return <WorkspacesLayout isAppBarHovered={isAppBarHovered} />;
+export function Workspaces() {
+  return <WorkspacesLayout />;
 }
