+++ packages/web-core/src/shared/components/ui-new/containers/SharedAppLayout.tsx
@@ -1,7 +1,6 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import type { DropResult } from '@hello-pangea/dnd';
 import { Outlet } from '@tanstack/react-router';
-import type { ReactNode } from 'react';
 import { siDiscord, siGithub } from 'simple-icons';
 import { XIcon, PlusIcon, LayoutIcon, KanbanIcon } from '@phosphor-icons/react';
 import { SyncErrorProvider } from '@/shared/providers/SyncErrorProvider';
@@ -36,6 +35,7 @@ import {
 import { OAuthDialog } from '@/shared/dialogs/global/OAuthDialog';
 import { CommandBarDialog } from '@/shared/dialogs/command-bar/CommandBarDialog';
 import { useCommandBarShortcut } from '@/shared/hooks/useCommandBarShortcut';
+import { useWorkspaceSidebarPreviewController } from '@/shared/hooks/useWorkspaceSidebarPreviewController';
 import { useShape } from '@/shared/integrations/electric/hooks';
 import { sortProjectsByOrder } from '@/shared/lib/projectOrder';
 import {
@@ -43,31 +43,24 @@ import {
   PROJECTS_SHAPE,
   type Project as RemoteProject,
 } from 'shared/remote-types';
+import { WorkspacesSidebarContainer } from '@/pages/workspaces/WorkspacesSidebarContainer';
+import { WorkspacesSidebarReopenTag } from '@vibe/ui/components/WorkspacesSidebar';
 
-interface AppBarHoverHandlers {
-  onHoverStart: () => void;
-  onHoverEnd: () => void;
-}
-
-interface SharedAppLayoutProps {
-  appBarHoverHandlers?: AppBarHoverHandlers;
-  children?: ReactNode;
-}
-
-export function SharedAppLayout({
-  appBarHoverHandlers,
-  children,
-}: SharedAppLayoutProps = {}) {
+export function SharedAppLayout() {
   const appNavigation = useAppNavigation();
   const currentDestination = useCurrentAppDestination();
   const isMigrateRoute = currentDestination?.kind === 'migrate';
   const isMobile = useIsMobile();
   const mobileFontScale = useUiPreferencesStore((s) => s.mobileFontScale);
+  const isLeftSidebarVisible = useUiPreferencesStore(
+    (s) => s.isLeftSidebarVisible
+  );
   const { isSignedIn } = useAuth();
   const { appVersion } = useUserSystem();
   const { data: onlineCount } = useDiscordOnlineCount();
   const { data: starCount } = useGitHubStars();
   const [isDrawerOpen, setIsDrawerOpen] = useState(false);
+  const [isAppBarHovered, setIsAppBarHovered] = useState(false);
 
   // Register CMD+K shortcut globally for all routes under SharedAppLayout
   useCommandBarShortcut(() => CommandBarDialog.show());
@@ -171,7 +164,13 @@ export function SharedAppLayout({
     [currentDestination]
   );
   const isWorkspacesActive = isWorkspacesDestination(currentDestination);
+  const isWorkspaceSidebarPreviewEnabled =
+    !isMobile && isWorkspacesActive && !isLeftSidebarVisible;
   const activeProjectId = projectDestination?.projectId ?? null;
+  const sidebarPreview = useWorkspaceSidebarPreviewController({
+    enabled: isWorkspaceSidebarPreviewEnabled,
+    isAppBarHovered,
+  });
 
   // Persist last selected project to scratch store
   const setSelectedProjectId = useUiPreferencesStore(
@@ -307,8 +306,8 @@ export function SharedAppLayout({
             isLoadingProjects={isLoading}
             onSignIn={handleSignIn}
             onMigrate={handleMigrate}
-            onHoverStart={appBarHoverHandlers?.onHoverStart}
-            onHoverEnd={appBarHoverHandlers?.onHoverEnd}
+            onHoverStart={() => setIsAppBarHovered(true)}
+            onHoverEnd={() => setIsAppBarHovered(false)}
             userPopover={
               <AppBarUserPopoverContainer
                 organizations={organizations}
@@ -451,8 +450,36 @@ export function SharedAppLayout({
             onOrgSelect={setSelectedOrgId}
             onOpenDrawer={() => setIsDrawerOpen(true)}
           />
-          <div className="flex-1 min-h-0 overflow-hidden">
-            {children ?? <Outlet />}
+          <div className="relative flex-1 min-h-0 overflow-hidden">
+            {isWorkspaceSidebarPreviewEnabled && (
+              <div className="absolute inset-y-0 left-0 z-20 flex items-center">
+                <WorkspacesSidebarReopenTag
+                  active={sidebarPreview.isPreviewOpen}
+                  onHoverStart={sidebarPreview.handleHandleHoverStart}
+                  onHoverEnd={sidebarPreview.handleHandleHoverEnd}
+                  ariaLabel="Workspaces"
+                />
+              </div>
+            )}
+
+            {isWorkspaceSidebarPreviewEnabled && (
+              <div
+                className={cn(
+                  'absolute left-0 top-0 z-30 h-full w-[300px] transition-transform duration-150 ease-out',
+                  sidebarPreview.isPreviewOpen
+                    ? 'translate-x-0 pointer-events-auto'
+                    : '-translate-x-full pointer-events-none'
+                )}
+                onMouseEnter={sidebarPreview.handlePreviewHoverStart}
+                onMouseLeave={sidebarPreview.handlePreviewHoverEnd}
+              >
+                <div className="h-full w-full overflow-hidden border-r border-border bg-secondary shadow-lg">
+                  <WorkspacesSidebarContainer />
+                </div>
+              </div>
+            )}
+
+            <Outlet />
           </div>
         </div>
       </div>
